setup(flavor): environment dev and prod
