+++ pubspec.yaml
@@ -7,6 +7,8 @@ environment:
   sdk: '>=2.19.3 <3.0.0'
 
 dependencies:
+  #ENV
+  flavor: ^2.0.0
   flutter:
     sdk: flutter
 
@@ -17,7 +19,6 @@ dependencies:
   #API 
   json_annotation: ^4.8.0
   json_serializable: ^6.6.0
-
 
 dev_dependencies:
   build_runner: ^2.3.3
